+++ packages/eas-cli/src/credentials/context.ts
@@ -1,4 +1,5 @@
 import { ExpoConfig, getConfig } from '@expo/config';
+import { Env } from '@expo/eas-build-job';
 import chalk from 'chalk';
 
 import Log from '../log';
@@ -27,6 +28,7 @@ export class CredentialsContext {
       nonInteractive?: boolean;
       projectDir: string;
       user: Actor;
+      env?: Env;
     }
   ) {
     this.projectDir = options.projectDir;
@@ -35,11 +37,21 @@ export class CredentialsContext {
 
     this.resolvedExp = options.exp;
     if (!this.resolvedExp) {
-      try {
-        this.resolvedExp = getExpoConfig(options.projectDir);
-      } catch (error) {
-        // ignore error, context might be created outside of expo project
-      }
+      this.resolvedExp =
+        CredentialsContext.getExpoConfigInProject(this.projectDir, { env: options.env }) ??
+        undefined;
+    }
+  }
+
+  static getExpoConfigInProject(
+    projectDir: string,
+    { env }: { env?: Env } = {}
+  ): ExpoConfig | null {
+    try {
+      return getExpoConfig(projectDir, { env });
+    } catch (error) {
+      // ignore error, context might be created outside of expo project
+      return null;
     }
   }