+++ packages/eas-cli/src/credentials/manager/ManageAndroid.ts
@@ -10,7 +10,7 @@ import {
 } from '../../project/projectUtils';
 import { promptAsync } from '../../prompts';
 import { findAccountByName } from '../../user/Account';
-import { ensureActorHasUsername } from '../../user/actions';
+import { ensureActorHasUsername, ensureLoggedInAsync } from '../../user/actions';
 import { AssignFcm } from '../android/actions/AssignFcm';
 import { AssignGoogleServiceAccountKey } from '../android/actions/AssignGoogleServiceAccountKey';
 import {
@@ -51,12 +51,18 @@ import {
 import { SelectBuildProfileFromEasJson } from './SelectBuildProfileFromEasJson';
 
 export class ManageAndroid {
-  constructor(private callingAction: Action) {}
+  constructor(private callingAction: Action, private projectDir: string) {}
 
-  async runAsync(
-    ctx: CredentialsContext,
-    currentActions: ActionInfo[] = highLevelActions
-  ): Promise<void> {
+  async runAsync(currentActions: ActionInfo[] = highLevelActions): Promise<void> {
+    const hasProjectContext = !!CredentialsContext.getExpoConfigInProject(this.projectDir);
+    const buildProfile = hasProjectContext
+      ? await new SelectBuildProfileFromEasJson(this.projectDir, Platform.ANDROID).runAsync()
+      : null;
+    const ctx = new CredentialsContext({
+      projectDir: process.cwd(),
+      user: await ensureLoggedInAsync(),
+      env: buildProfile?.env,
+    });
     const accountName = ctx.hasProjectContext
       ? getProjectAccountName(ctx.exp, ctx.user)
       : ensureActorHasUsername(ctx.user);
@@ -64,7 +70,12 @@ export class ManageAndroid {
     if (!account) {
       throw new Error(`You do not have access to account: ${accountName}`);
     }
-    const { gradleContext } = await this.createProjectContextAsync(ctx);
+
+    let gradleContext;
+    if (ctx.hasProjectContext) {
+      assert(buildProfile, 'buildProfile must be defined in a project context');
+      gradleContext = await this.createProjectContextAsync(ctx, buildProfile);
+    }
 
     while (true) {
       try {
@@ -140,30 +151,18 @@ export class ManageAndroid {
     }
   }
 
-  private async createProjectContextAsync(ctx: CredentialsContext): Promise<{
-    gradleContext?: GradleBuildContext;
-    buildProfile?: BuildProfile<Platform.ANDROID>;
-  }> {
-    if (!ctx.hasProjectContext) {
-      return {};
-    }
-
+  private async createProjectContextAsync(
+    ctx: CredentialsContext,
+    buildProfile: BuildProfile<Platform.ANDROID>
+  ): Promise<GradleBuildContext | undefined> {
+    assert(ctx.hasProjectContext, 'createProjectContextAsync: must have project context.');
     const maybeProjectId = await promptToCreateProjectIfNotExistsAsync(ctx.exp);
     if (!maybeProjectId) {
       throw new Error(
         'Your project must be registered with EAS in order to use the credentials manager.'
       );
     }
-
-    const buildProfile = await new SelectBuildProfileFromEasJson(
-      ctx.projectDir,
-      Platform.ANDROID
-    ).runAsync(ctx);
-    const gradleContext = await resolveGradleBuildContextAsync(ctx.projectDir, buildProfile);
-    return {
-      gradleContext,
-      buildProfile,
-    };
+    return await resolveGradleBuildContextAsync(ctx.projectDir, buildProfile);
   }
 
   private async runProjectSpecificActionAsync(