+++ packages/eas-cli/src/credentials/manager/SelectPlatform.ts
@@ -1,10 +1,9 @@
 import { promptAsync } from '../../prompts';
-import { CredentialsContext } from '../context';
 import { ManageAndroid } from './ManageAndroid';
 import { ManageIos } from './ManageIos';
 
 export class SelectPlatform {
-  async runAsync(ctx: CredentialsContext): Promise<void> {
+  async runAsync(): Promise<void> {
     const { platform } = await promptAsync({
       type: 'select',
       name: 'platform',
@@ -16,8 +15,8 @@ export class SelectPlatform {
     });
 
     if (platform === 'ios') {
-      return await new ManageIos(new SelectPlatform()).runAsync(ctx);
+      return await new ManageIos(new SelectPlatform(), process.cwd()).runAsync();
     }
-    return await new ManageAndroid(new SelectPlatform()).runAsync(ctx);
+    return await new ManageAndroid(new SelectPlatform(), process.cwd()).runAsync();
   }
 }